Transform and build charm from ``ovn-chassis`` repo
The ``ovn-dedicated-chassis`` has the exact same code base as its
sibling ``ovn-chassis``, the only difference being its name and
status as a principle charm as opposed to subordinate charm.

Add script to transform ``ovn-chassis`` into ``ovn-dedicated-chassis``
along with a tox.ini with modified build step to make it all work.

Change-Id: Ib133c805151cf66c5316df38c33ea6d20a38fa2b