Rebuild to pull in updated sriov-netplan-shim
The ovn-dedicated-chassis charm consumes the sriov-netplan-shim
Python package as a wheel for its `pci` module which is used to
render the `/etc/sriov-netplan-shim/interfaces.yaml` configuration
file used for boot time configuration of SR-IOV VFs.

The package was recently updated and the chages are required for
the charm to correctly interpret sysfs information when used with
recent versions of the mlx5_core driver.

Related-Bug: #1892132
Related-Bug: #1927125
Change-Id: I17792b718024d6ff7ff52394fe33f06db971e8b1