Default to OVN 22.03 for fresh deployments on Focal
Add `ovn-source` configuration option which should be used in
addition to the `source` configuration option.

The default of the `ovn-source` configuration option is
determined at runtime based on whether the end user has provided
config, whether its a fresh or upgraded charm deployment and which
series it is deployed on.

Change-Id: I3d55bafb598339b0b69ad0d65473d79bf9cdbe4c