Add docs key and point at Discourse
Add the 'docs' key and point it at a Discourse topic
previously populated with the charm's README contents.

When the new charm revision is released to the Charmhub,
this Discourse-based content will be displayed there. In
the absense of the this new key, the Charmhub's default
behaviour is to display the value of the charm's
'description' key.

Add charmhub-pop topic

Change-Id: I7ac6283884e85c319a634e210da8972d49abff8c
(cherry picked from commit 1090b6e196f38f54c61cd13221da40941b4b8fb6)